rendering ListStats in Home
